reuse deDupeThings for has_components
