+++ src/components/TextListWithHeading.js
@@ -2,6 +2,7 @@ import React from "react";
 import { FormattedMessage } from "react-intl";
 import _ from "underscore";
 import "./ListWithHeading.css";
+import { deDupeThings } from "../util.js";
 
 export default class TextListGroupWithHeading extends React.Component {
   render() {
@@ -11,9 +12,12 @@ export default class TextListGroupWithHeading extends React.Component {
     }
     let value = thing[property];
     if (!value || (_.isArray(value) && !value.length)) {
-      return <div />;
-    } else if (_.isArray(thing[property])) {
-      let items = thing[property].map(item => (
+      return null;
+    } else if (_.isArray(value)) {
+      if (property === "has_components") {
+        value = deDupeThings(value);
+      }
+      let items = value.map(item => (
         <div key={item.value}>
           <FormattedMessage id={item.text || item.title || item} />
         </div>
